+++ OMG.PSUtilities.AI/Public/Invoke-PSUGitCommit.ps1
@@ -225,8 +225,8 @@ $($item.Diff)
         Write-Host "Following is the Commit message!" -ForegroundColor Cyan
         Write-Host $CommitMessage -ForegroundColor DarkYellow
 
-        Write-Host "[R]      --> Regenerate a new commit message!" -ForegroundColor Cyan
-        Write-Host "`n[Ctrl+C] --> Abort commit process!" -ForegroundColor Cyan
+        Write-Host "`n[R]      --> Regenerate a new commit message!" -ForegroundColor Cyan
+        Write-Host "[Ctrl+C] --> Abort commit process!" -ForegroundColor Cyan
         Write-Host "[Enter]  --> Accept the above commit message!" -ForegroundColor Cyan
 
         $CustomCommitMsg = Read-Host -Prompt "Enter your choice"